requirements.txt
openai==1.10.0
python-dotenv==1.0.0
requests==2.31.0
streamlit==1.32.0
pandas==2.0.3
matplotlib==3.7.2